# Device enumeration perf