--- a/streamdeck-audiometer/Actions/AudioMeterAction.cs
+++ b/streamdeck-audiometer/Actions/AudioMeterAction.cs
@@ -12,6 +12,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Timers;
 
 namespace AudioMeter.Actions
 {
@@ -253,7 +254,14 @@ namespace AudioMeter.Actions
         {
             if (mmDevice == null)
             {
+                if (!(sender is Timer timer)) return;
+                var interval = timer.Interval;
+
+                timer.Interval = 30000;
+
                 SetMMDeviceFromDeviceName(settings.AudioDevice);
+
+                timer.Interval = Math.Min(200, interval);
             }
 
             if (mmDevice != null)
@@ -283,7 +291,7 @@ namespace AudioMeter.Actions
             }
 
             MMDeviceEnumerator enumerator = new MMDeviceEnumerator();
-            mmDevice = enumerator.EnumerateAudioEndPoints(DataFlow.All, DeviceState.Active).ToList().Where(d => d.FriendlyName == deviceName).FirstOrDefault();
+            mmDevice = enumerator.EnumerateAudioEndPoints(DataFlow.All, DeviceState.Active).ToList().FirstOrDefault(d => d.FriendlyName == deviceName);
         }
 
         private async Task DisplayMeter(int level)