--- a/streamdeck-audiometer/packages.config
+++ b/streamdeck-audiometer/packages.config
@@ -1,19 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="CommandLineParser" version="2.8.0" targetFramework="net472" />
-  <package id="Microsoft.Win32.Registry" version="5.0.0" targetFramework="net472" />
-  <package id="NAudio" version="2.0.0" targetFramework="net472" />
-  <package id="NAudio.Asio" version="2.0.0" targetFramework="net472" />
-  <package id="NAudio.Core" version="2.0.0" targetFramework="net472" />
-  <package id="NAudio.Midi" version="2.0.0" targetFramework="net472" />
-  <package id="NAudio.Wasapi" version="2.0.0" targetFramework="net472" />
-  <package id="NAudio.WinForms" version="2.0.0" targetFramework="net472" />
-  <package id="NAudio.WinMM" version="2.0.0" targetFramework="net472" />
+  <package id="CommandLineParser" version="2.9.1" targetFramework="net472" />
+  <package id="NAudio" version="2.1.0" targetFramework="net472" />
+  <package id="NAudio.Asio" version="2.1.0" targetFramework="net472" />
+  <package id="NAudio.Core" version="2.1.0" targetFramework="net472" />
+  <package id="NAudio.Midi" version="2.1.0" targetFramework="net472" />
+  <package id="NAudio.Wasapi" version="2.1.0" targetFramework="net472" />
+  <package id="NAudio.WinForms" version="2.1.0" targetFramework="net472" />
+  <package id="NAudio.WinMM" version="2.1.0" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
-  <package id="NLog" version="4.7.10" targetFramework="net472" />
+  <package id="NLog" version="5.0.1" targetFramework="net472" />
   <package id="streamdeck-client-csharp" version="4.3.0" targetFramework="net472" />
   <package id="StreamDeck-Tools" version="3.2.0" targetFramework="net472" />
-  <package id="System.Drawing.Common" version="5.0.2" targetFramework="net472" />
-  <package id="System.Security.AccessControl" version="5.0.0" targetFramework="net472" />
-  <package id="System.Security.Principal.Windows" version="5.0.0" targetFramework="net472" />
+  <package id="System.Drawing.Common" version="6.0.0" targetFramework="net472" />
 </packages>