--- a/streamdeck-audiometer/packages.config
+++ b/streamdeck-audiometer/packages.config
@@ -9,7 +9,7 @@
   <package id="NAudio.WinForms" version="2.1.0" targetFramework="net472" />
   <package id="NAudio.WinMM" version="2.1.0" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
-  <package id="NLog" version="5.0.1" targetFramework="net472" />
+  <package id="NLog" version="4.7.6" targetFramework="net472" />
   <package id="streamdeck-client-csharp" version="4.3.0" targetFramework="net472" />
   <package id="StreamDeck-Tools" version="3.2.0" targetFramework="net472" />
   <package id="System.Drawing.Common" version="6.0.0" targetFramework="net472" />