Steps to reproduce:
- Have many (10-15+) audio devices to enumerate
- Place a vu meter for a device onto 2 different stream deck pages
- Watch the task manager in Windows for the audiometer sub-process and cpu usage when starting stream deck software up
- Watch the task manager again when switching pages in the stream deck (causes tear down / dispose and re-initialization of button)

Expected results:
- Enumeration to complete quickly and VU meter to display valid numbers

Actual results:
- Enumeration takes upwards of 20s to complete and a cpu core is maxed the whole time, display produces nothing until this completes

Root cause:
- Although the event for the display update is an async method, it's called directly and on a 200ms repeat timer in separate threads by the caller; the wrapper sdk
- This means if it takes 2s to enumerate devices once, by the time 2s has passed, there's already 10 other threads all trying to enumerate the same data in other threads concurrently
- As this all takes place on a single core, it just locks it up until all the threads finish enumerating, around 20s give or take

Solution:
- Since we don't have direct control over the caller to stop it invoking multiple times, we can slow things down ourselves
- This also makes the async approach moot because there's no awaiting/callback for the results anyway, it's synchronously called, plus it wouldn't stop a separate async call in each thread from all enumerating on a line
- So instead we take over the interval of the timer on the first call to us (when current device is null and enumeration is about to start) and ramp it up to 30s
- Enumeration then takes place as normal, but only takes 2s
- During this time, the timer doesn't spin up any additional threads
- Once enumeration completes, we ramp down the timer interval to 200ms or its prior value to us, whichever was lower (it is currently defaulting to 200ms anyway)
- Subsequent calls then operate every ~200ms as normal and skip further enumeration and just grab the value they want to display