@ndarlington 2022-07-21
Note: We're running with updated nuget package references that our fork master isn't using. I tried just pulling in the existing packages at the solution's stated versions, and it ran into many issues trying to resolve the references to them. Most I had to fully remove from the solution and re-add to resolve all the problems and get it building.